pyproject.toml
[project]
name = "memgen"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.40",
    "datasets>=3.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "pyyaml>=6.0",
    "click>=8.0",
    "tenacity>=8.0",
    "tqdm>=4.0",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-asyncio>=0.23"]

[project.scripts]
memgen = "memgen.cli:main"

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
